requirements.txt
requests
imapclient
mail-parser
PyPDF2
python-docx
mailjet_rest
python-dotenv